🎹 Visible particles Restructured Temporarily removed all multithreading
