+++ src/actions/play.rs
@@ -7,14 +7,15 @@ use crate::Pianorium;
 
 impl Pianorium {
     pub fn play(&mut self) -> Result<(), String> {
+        println!("Playing the visualization…");
+        self.ogl.to_zero();
         self.winsdl.window
             .subsystem()
             .gl_set_swap_interval(SwapInterval::VSync)
             .unwrap();
 
-        let ogl = self.handles.remove(0).join().unwrap();
-        let mut play_ogl = ogl.clone();
-        play_ogl.to_zero();
+        // let ogl = self.handles.remove(0).join().unwrap();
+        // self.ogl.to_zero();
         
         let start_time = Instant::now();
         let mut since_last: f32;
@@ -29,15 +30,16 @@ impl Pianorium {
             since_last = start_time.elapsed().as_secs_f32()-since_start;
             since_start += since_last;
 
-            if play_ogl.frame > play_ogl.shared.max_frame { break 'play; }                                                                    // Stop when it's finished playing
+            if self.ogl.frame > self.ogl.max_frame { break 'play; }                                                                    // Stop when it's finished playing
             
-            unsafe { gl::Uniform1f(play_ogl.shared.u_time.id, since_start as f32); }  
-            play_ogl.draw();
-            play_ogl.update(since_last);
+            unsafe { gl::Uniform1f(self.ogl.u_time.id, since_start as f32); }  
+            self.ogl.update(since_last);
+            self.ogl.draw();
+            self.ogl.frame += 1;
             self.winsdl.window.gl_swap_window();
 
         }
-        self.handles.insert(0, std::thread::spawn(move ||{ ogl }));
+        // self.handles.insert(0, std::thread::spawn(move ||{ ogl }));
 
         Ok(())
     }
