+++ src/main.rs
@@ -4,7 +4,6 @@ extern crate num_cpus;
 pub mod actions;
 pub mod fs;
 pub mod opengl;
-pub mod midi;
 pub mod pianorium;
 pub mod render;
 pub mod ui;
@@ -12,8 +11,7 @@ pub mod window;
 
 pub use actions::capture::*;
 pub use fs::*;
-pub use opengl::{context::*, drawing::*, shaders::*, uniforms::*};
-pub use midi::*;
+pub use opengl::{context::*, drawing::*, layout::*, notes::*, particles::*, shaders::*, uniforms::*};
 pub use pianorium::*;
 pub use render::*;
 pub use ui::*;
@@ -22,8 +20,8 @@ pub use window::*;
 fn main() {
     let mut p = pianorium::Pianorium::new().unwrap();
     p.play().unwrap();
-    p.full_png().unwrap();
     p.full_mp4().unwrap();
+    p.full_png().unwrap();
 }
 
 // fn draw_gui() { // Struct with Impl
