+++ src/opengl/mod.rs
@@ -1,9 +1,15 @@
 pub mod context;
 pub mod drawing;
+pub mod layout;
+pub mod notes;
+pub mod particles;
 pub mod shaders;
 pub mod uniforms;
 
 pub use context::*;
 pub use drawing::*;
+pub use layout::*;
+pub use notes::*;
+pub use particles::*;
 pub use shaders::*;
 pub use uniforms::*;
