+++ src/render/mod.rs
@@ -2,15 +2,17 @@ use std::{process::{Command, Stdio}, io::Write};
 
 use egui_sdl2_gl::gl;
 
+use crate::OpenGLContext;
 
-impl crate::opengl::context::OpenGLContext {
+
+impl OpenGLContext {
     pub fn read(&mut self) {
         unsafe {
             gl::ReadPixels(
                 0,
                 0,
-                self.shared.width as i32,
-                self.shared.height as i32,
+                self.width as i32,
+                self.height as i32,
                 gl::BGRA,
                 gl::UNSIGNED_BYTE,
                 self.data.as_mut_ptr() as *mut gl::types::GLvoid,
@@ -26,9 +28,9 @@ impl crate::opengl::context::OpenGLContext {
             .env("FFREPORT", "file=ff_export_mp4.log:level=56")
             .arg("-loglevel").arg("0")
             .arg("-f").arg("rawvideo")
-            .arg("-r").arg(self.shared.framerate.to_string())
+            .arg("-r").arg(self.framerate.to_string())
             .arg("-pix_fmt").arg("bgra")
-            .arg("-s").arg(format!("{}x{}", self.shared.width, self.shared.height))
+            .arg("-s").arg(format!("{}x{}", self.width, self.height))
             .arg("-i").arg("-")
             .arg("-vcodec").arg("libx264")
             .arg("-crf").arg("23")
@@ -49,7 +51,7 @@ impl crate::opengl::context::OpenGLContext {
             .arg("-loglevel").arg("0")
             .arg("-f").arg("rawvideo")
             .arg("-pix_fmt").arg("bgra")
-            .arg("-s").arg(format!("{}x{}", self.shared.width, self.shared.height))
+            .arg("-s").arg(format!("{}x{}", self.width, self.height))
             .arg("-i").arg("-")
             .arg("-frames:v").arg("1")
             .arg("-vf").arg("vflip")
